backend/requirements.txt
fastapi
uvicorn
pymongo
spacy
nltk
sentence-transformers
python-multipart
pydantic-settings
python-dotenv
PyMuPDF
python-docx
motor
